+++ src/DefaultDomains.php
@@ -1,0 +1,173 @@
+<?php
+
+namespace MischaSigtermans\Sift;
+
+class DefaultDomains
+{
+    /**
+     * Default list of 100+ public email providers.
+     *
+     * Organized by category and sorted alphabetically within each.
+     *
+     * @var array<string>
+     */
+    public const LIST = [
+
+        // Global providers
+        'aol.com',
+        'gmail.com',
+        'googlemail.com',
+        'hotmail.co.uk',
+        'hotmail.com',
+        'hotmail.de',
+        'hotmail.es',
+        'hotmail.fr',
+        'hotmail.it',
+        'icloud.com',
+        'live.co.uk',
+        'live.com',
+        'live.de',
+        'live.es',
+        'live.fr',
+        'live.it',
+        'mac.com',
+        'me.com',
+        'msn.com',
+        'outlook.co.uk',
+        'outlook.com',
+        'outlook.de',
+        'outlook.es',
+        'outlook.fr',
+        'outlook.it',
+        'yahoo.ca',
+        'yahoo.co.in',
+        'yahoo.co.jp',
+        'yahoo.co.uk',
+        'yahoo.com',
+        'yahoo.com.au',
+        'yahoo.com.br',
+        'yahoo.com.sg',
+        'yahoo.de',
+        'yahoo.es',
+        'yahoo.fr',
+        'yahoo.it',
+
+        // Privacy-focused
+        'countermail.com',
+        'disroot.org',
+        'fastmail.com',
+        'hushmail.com',
+        'mailfence.com',
+        'posteo.de',
+        'proton.me',
+        'protonmail.ch',
+        'protonmail.com',
+        'tutanota.com',
+
+        // US ISPs & telecom
+        'att.net',
+        'bellsouth.net',
+        'charter.net',
+        'comcast.net',
+        'cox.net',
+        'earthlink.net',
+        'juno.com',
+        'netscape.net',
+        'sbcglobal.net',
+        'verizon.net',
+        'windstream.net',
+
+        // Regional - Germany
+        'freenet.de',
+        'gmx.at',
+        'gmx.ch',
+        'gmx.com',
+        'gmx.de',
+        'gmx.net',
+        't-online.de',
+        'web.de',
+
+        // Regional - France
+        'free.fr',
+        'laposte.net',
+        'orange.fr',
+
+        // Regional - Italy
+        'alice.it',
+        'libero.it',
+        'tiscali.it',
+        'virgilio.it',
+
+        // Regional - Spain
+        'telefonica.net',
+
+        // Regional - Netherlands
+        'kpnmail.nl',
+        'ziggo.nl',
+
+        // Regional - Switzerland
+        'bluewin.ch',
+
+        // Regional - Czech Republic
+        'centrum.cz',
+        'seznam.cz',
+
+        // Regional - Russia & Eastern Europe
+        'bigmir.net',
+        'bk.ru',
+        'inbox.ru',
+        'list.ru',
+        'mail.bg',
+        'mail.ru',
+        'rambler.ru',
+        'yandex.com',
+        'yandex.ru',
+
+        // Regional - Asia
+        '126.com',
+        '163.com',
+        'docomo.ne.jp',
+        'ezweb.ne.jp',
+        'naver.com',
+        'qq.com',
+        'sina.com',
+        'softbank.ne.jp',
+        'sohu.com',
+
+        // Regional - Brazil
+        'bol.com.br',
+        'terra.com.br',
+        'uol.com.br',
+
+        // Regional - Other
+        'bigpond.com',
+        'indiatimes.com',
+        'rediffmail.com',
+
+        // General providers
+        'email.com',
+        'excite.com',
+        'lycos.com',
+        'mail.com',
+        'rocketmail.com',
+        'zoho.com',
+        'zoho.com.au',
+        'zoho.eu',
+        'zoho.in',
+
+        // Temporary & disposable
+        '10minutemail.com',
+        'burnermail.io',
+        'fakeinbox.com',
+        'getnada.com',
+        'guerrillamail.com',
+        'maildrop.cc',
+        'mailinator.com',
+        'temp-mail.org',
+        'tempmail.com',
+        'throwaway.email',
+        'trashmail.com',
+        'yopmail.com',
+
+    ];
+}